refactor audio: move local service/history into music namespace
