+++ src/yoyopod/audio/__init__.py
@@ -1,15 +1,17 @@
 """App-facing seam for the media/audio domain."""
 
-from yoyopod.audio.history import RecentTrackEntry, RecentTrackHistoryStore
-from yoyopod.audio.local_service import LocalLibraryItem, LocalMusicService
-from yoyopod.audio.manager import AudioManager, AudioDevice
+from yoyopod.audio.manager import AudioDevice, AudioManager
 from yoyopod.audio.music import (
-    MusicBackend,
+    LocalLibraryItem,
+    LocalMusicService,
     MockMusicBackend,
     MpvBackend,
+    MusicBackend,
     MusicConfig,
     PlaybackQueue,
     Playlist,
+    RecentTrackEntry,
+    RecentTrackHistoryStore,
     Track,
 )
 from yoyopod.audio.volume import OutputVolumeController
