+++ src/yoyopod/audio/music/__init__.py
@@ -2,6 +2,8 @@
 
 from yoyopod.audio.music.backend import MockMusicBackend, MpvBackend, MusicBackend
 from yoyopod.audio.music.models import MusicConfig, PlaybackQueue, Playlist, Track
+from yoyopod.audio.music.history import RecentTrackEntry, RecentTrackHistoryStore
+from yoyopod.audio.music.library import LocalLibraryItem, LocalMusicService
 
 __all__ = [
     "MockMusicBackend",
@@ -10,5 +12,9 @@ __all__ = [
     "MusicConfig",
     "PlaybackQueue",
     "Playlist",
+    "LocalLibraryItem",
+    "LocalMusicService",
+    "RecentTrackEntry",
+    "RecentTrackHistoryStore",
     "Track",
 ]
