+++ src/yoyopod/coordinators/playback.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 from loguru import logger
 
-from yoyopod.audio.local_service import LocalMusicService
+from yoyopod.audio.music import LocalMusicService
 from yoyopod.audio.music.models import Track
 from yoyopod.coordinators.runtime import AppRuntimeState, CoordinatorRuntime
 from yoyopod.coordinators.screen import ScreenCoordinator
